+++ services/src/main/java/org/keycloak/authentication/authenticators/broker/IdpReviewProfileAuthenticatorFactory.java
@@ -39,6 +39,7 @@ public class IdpReviewProfileAuthenticatorFactory implements AuthenticatorFactor
     static IdpReviewProfileAuthenticator SINGLETON = new IdpReviewProfileAuthenticator();
 
     public static final String UPDATE_PROFILE_ON_FIRST_LOGIN = "update.profile.on.first.login";
+    public static final String TERMS_AND_CONDITIONS = "terms_and_conditions";
 
     @Override
     public Authenticator create(KeycloakSession session) {
@@ -110,7 +111,12 @@ public class IdpReviewProfileAuthenticatorFactory implements AuthenticatorFactor
                 + " page for reviewing profile will be displayed and user can review and update his profile. Value 'off' means that page won't be displayed."
                 + " Value 'missing' means that page is displayed just when some required attribute is missing (wasn't downloaded from identity provider). Value 'missing' is the default one."
                 + " WARN: In case that user clicks 'Review profile info' on link duplications page, the update page will be always displayed. You would need to disable this authenticator to never display the page.");
-
+        configProperties.add(property);
+        property = new ProviderConfigProperty();
+        property.setName(TERMS_AND_CONDITIONS);
+        property.setLabel("Accept Terms and Conditions");
+        property.setType(ProviderConfigProperty.BOOLEAN_TYPE);
+        property.setHelpText("If this option is true and terms and conditions required action is enabled, new user will be obliged to accept terms and conditions during profile review. With this first broken login flow will follow GDPR.");
         configProperties.add(property);
     }