+++ testsuite/integration-arquillian/tests/base/src/main/java/org/keycloak/testsuite/pages/LoginUpdateProfilePage.java
@@ -97,7 +97,7 @@ public class LoginUpdateProfilePage extends AbstractPage {
     }
 
     public boolean isDepartmentEnabled() {
-        return departmentInput.isEnabled();
+        return departmentInput.getAttribute("readOnly") == null;
     }
 
     public boolean isCurrent() {